[refactor] #50 domain repository local remote 삭제
