+++ app/src/main/java/org/android/bbangzip/data/repository/remote/DummyRepositoryImpl.kt
@@ -2,7 +2,7 @@ package org.android.bbangzip.data.repository.remote
 
 import org.android.bbangzip.data.source.remote.datasource.DummyRemoteDataSource
 import org.android.bbangzip.domain.model.Dummy
-import org.android.bbangzip.domain.repository.remote.DummyRepository
+import org.android.bbangzip.domain.repository.DummyRepository
 import javax.inject.Inject
 
 class DummyRepositoryImpl
