+++ app/src/main/java/org/android/bbangzip/data/repository/remote/TimerRepositoryImpl.kt
@@ -3,7 +3,7 @@ package org.android.bbangzip.data.repository.remote
 import org.android.bbangzip.data.source.remote.datasource.TimerRemoteDataSource
 import org.android.bbangzip.data.source.remote.dto.request.RequestCompleteTimerDto
 import org.android.bbangzip.domain.model.BreadList
-import org.android.bbangzip.domain.repository.remote.TimerRepository
+import org.android.bbangzip.domain.repository.TimerRepository
 import javax.inject.Inject
 
 class TimerRepositoryImpl
