+++ app/src/main/java/org/android/bbangzip/presentation/ui/dummy/DummyViewModel.kt
@@ -7,7 +7,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import org.android.bbangzip.UserPreferences
-import org.android.bbangzip.domain.repository.local.UserRepository
+import org.android.bbangzip.domain.repository.UserRepository
 import org.android.bbangzip.domain.usecase.FetchDummyUseCase
 import org.android.bbangzip.presentation.common.base.BaseViewModel
 import javax.inject.Inject
